+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.0b4 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+- Remove unused newsview div and rename mediabox with galleria div.
+  [marcosfromero]
+
 - Clean and add styles.css in cssregistry. [marcosfromero]
 
 - Rename nitf_galleria template with galleria. [marcosfromero]